--- tests/implementations/models.py
@@ -1,13 +1,0 @@
-from tortoise import Model, fields
-
-
-class PotatoModel(Model):
-    thickness = fields.FloatField()
-    mass = fields.FloatField()
-    color = fields.CharField(max_length=255)
-    type = fields.CharField(max_length=255)
-
-
-class CarrotModel(Model):
-    length = fields.FloatField()
-    color = fields.CharField(max_length=255)
